+++ CHANGES.rst
@@ -7,6 +7,7 @@ The ASDF Standard is at v1.6.0
 - Add AsdfProvisionalAPIWarning to warn developers of new features that
   may undergo breaking changes but are likely to be included as stable
   features (without this warning) in a future version of ASDF [#1295]
+- Add AsdfDeprecationWarning to AsdfFile.blocks [#1336]
 
 2.14.3 (2022-12-15)
 -------------------